makeTheDreamWork(): Re-structured exception management to be identified as early as in run_shell_process
This is to enable a more generic approach to exception management. In this case we can also categorize errors coming from the base CommandLine process as their own, and if nothing matches we can throw the entire error back to the caller, letting them implement their own type then and there.