wingetnpm(): cant cache exceptions in static variable
Because of the way Pythons runtime works it won't have access to the entire assembly when we ask for all subclasses from a class. Because of this we need to make this scan every time we do a exception lookup. Less performant but will act more as expected.

Changed the name of NpmError to NpmException